Completed create a spot endpoint

+++ backend/db/models/spot.js
@@ -56,12 +56,20 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false
     },
     lat: {
-      type: DataTypes.DECIMAL,
-      allowNull: false
+      type: DataTypes.FLOAT,
+      allowNull: false,
+      // validate: {
+      //   min: -90,
+      //   max: 90
+      // }
     },
     lng: {
-      type: DataTypes.DECIMAL,
-      allowNull: false
+      type: DataTypes.FLOAT,
+      allowNull: false,
+      // validate: {
+      //   min: -180,
+      //   max: 180
+      // }
     },
     name: {
       type: DataTypes.STRING,
@@ -72,7 +80,7 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false
     },
     price: {
-      type: DataTypes.DECIMAL,
+      type: DataTypes.FLOAT,
       allowNull: false,
       validate: {
         min: 0
